fixed more time truncation errors

+++ anc_tools/utils/serial2epoch.m
@@ -54,6 +54,6 @@ if (serial < DAYS_TO_1970 | serial > DAYS_FROM_1970)
 end
 
 % Convert to epoch.
-epoch = (serial - DAYS_TO_1970) * 86400.0;
+epoch = double(serial - DAYS_TO_1970) * 86400.0;
 
 return
